% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkBarbieDimensions}
\alias{checkBarbieDimensions}
\title{check whether components of Barbie object have identical sample dimensions and Barcode dimensions.}
\usage{
checkBarbieDimensions(Barbie)
}
\arguments{
\item{Barbie}{'Barbie' object created by createBarbie()}
}
\value{
a logical value
}
\description{
check whether components of Barbie object have identical sample dimensions and Barcode dimensions.
}
\examples{
checkBarbieDimensions(Barbie)
}
