% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{collapseBarcodes}
\alias{collapseBarcodes}
\title{collapse a 'Barbie' object by Barcode groups}
\usage{
collapseBarcodes(Barbie, groupArray, method = max)
}
\arguments{
\item{Barbie}{an object created by the 'createBarbie()' function in the 'Barbie' package}

\item{groupArray}{a vector indication the group of each sample}

\item{method}{a function choosing method for collapsing Barcode counts within Barcode groups}
}
\value{
a 'Barbie' object with samples collapsed by 'groupArray'
}
\description{
collapse a 'Barbie' object by Barcode groups
}
\examples{
myBarbie <- createBarbie(
  object=matrix(c(1:20), nrow=5, ncol=4),
  target=data.frame(fac=c(1, 1, 2, 2)))
collapseBarcodes(Barbie=myBarbie, groupArray=c(1,1,2,2,2))
}
