% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{collapseColumnsByArray}
\alias{collapseColumnsByArray}
\title{Title collapse a matrix by column groups}
\usage{
collapseColumnsByArray(mat, groupArray, method = mean)
}
\arguments{
\item{mat}{a matrix, data.frame, or vector}

\item{groupArray}{a vector indication the groups of each column in 'mat'}

\item{method}{a function indicating the method for collapsing each Barcodes in the subset matrix of each sample group}
}
\value{
a matrix, data.frame, or vector collapsed by column groups
}
\description{
Title collapse a matrix by column groups
}
\examples{
mymatrix <- matrix(1:25, nrow=5, ncol=5)
collapseColumnsByArray(mymatrix, groupArray=c(1,1,1,2,2), method=mean)
}
