% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{collapseRowsByArray}
\alias{collapseRowsByArray}
\title{collapse a matrix by column groups}
\usage{
collapseRowsByArray(mat, groupArray, method = max)
}
\arguments{
\item{mat}{a matrix, data.frame, or vector}

\item{groupArray}{a vector indication the groups of each column in 'mat'}

\item{method}{function indicating the method for collapsing each sample in the subset matrix of each Barcode group}
}
\value{
a matrix, data.frame, or vector collapsed by column groups
}
\description{
collapse a matrix by column groups
}
\examples{
mymatrix <- matrix(1:25, nrow=5, ncol=5)
collapseRowsByArray(mymatrix, groupArray=c(1,1,1,2,2), method=mean)
}
