% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{collapseSamples}
\alias{collapseSamples}
\title{Title collapse a 'Barbie' object by sample groups}
\usage{
collapseSamples(Barbie, groupArray)
}
\arguments{
\item{Barbie}{an object created by the 'createBarbie()' function in the 'Barbie' package}

\item{groupArray}{a vector indication the group of each sample}
}
\value{
a 'Barbie' object with samples collapsed by 'groupArray'
}
\description{
Title collapse a 'Barbie' object by sample groups
}
\examples{
myBarbie <- createBarbie(
  object=matrix(c(1:20), nrow=5, ncol=4),
  target=data.frame(fac=c(1, 1, 2, 2)))
collapseSamples(Barbie=myBarbie, groupArray=c(1,1,2,2))
}
