% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{combineBarbieQs}
\alias{combineBarbieQs}
\title{Merge several \code{barbieQ} objects into one object organized by samples}
\usage{
combineBarbieQs(...)
}
\arguments{
\item{...}{A list of \code{barbieQ} objects created by the \link{createBarbieQ} function.}
}
\value{
A \code{barbieQ} object including all the samples passed \code{barbieQ} objects.
}
\description{
\code{combineBarbieQs} merges multiple \code{barbieQ} objects, aligning their components
based on samples. Components such as \code{assay}, \code{metadata}, \code{proportion},
\code{CPM}, \code{occurrence}, \code{rank}, and \code{isTop$mat} are combined by binding their
columns (samples). The \code{isTop$vec} component is merged using a logical
OR (\code{||}) operator, while other inherited components are passed unchanged.
}
\note{
All input \code{barbieQ} objects must have the same number of Barcodes
(rows) in the same order.
}
\examples{
barbieQ1 <- createBarbieQ(
  object = matrix(seq_len(20), nrow = 5, ncol = 4),
  target = data.frame(
    a = c("a", "a", "b", "b"),
    b = c("x", "x", "x", "x")
  ),
  factorColors = list(
    a = c("a" = "#FFFFFF", "b" = "#CCCCFF"),
    b = c("x" = "#000099")
  )
)
barbieQ2 <- createBarbieQ(
  object = matrix(seq(21, 20), nrow = 5, ncol = 2),
  target = data.frame(
    a = c("a", "c"),
    c = c("u", "u")
  ),
  factorColors = list(
    a = c("a" = "#FFFFFF", "c" = "#000000"),
    c = c("u" = "#CCCCCC")
  )
)
barbieQ3 <- createBarbieQ(
  object = matrix(seq(31, 35), nrow = 5, ncol = 1),
  target = data.frame(b = c("x")),
  factorColors = list(b = c("x" = "#999999"))
)
combineBarbieQs(barbieQ1, barbieQ2, barbieQ3)
}
