% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBarbie.R
\name{createBarbie}
\alias{createBarbie}
\title{create an object-centred structure-a 'Barbie' object-based on the input Barcode count and other information.}
\usage{
createBarbie(object, target = NULL, factorColors = NULL)
}
\arguments{
\item{object}{a numeric matrix or equivalent; or a Barbie object or equivalent}

\item{target}{a matrix of sample conditions}

\item{factorColors}{a list containing colour palettes corresponding to the factors in the target parameter}
}
\value{
a Barbie object
}
\description{
\code{createBarbie}
}
\examples{
createBarbie(mymat)
createBarbie(mymat, mytarget)
createBarbie(mymat, mytarget, mycolors)

}
