% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extarctTargetsAndPrimaryFactor}
\alias{extarctTargetsAndPrimaryFactor}
\title{extracting targets and primary factor based on the imported arguments}
\usage{
extarctTargetsAndPrimaryFactor(Barbie, targets = NULL, groupBy = NULL)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{targets}{a matrix or data.frame storing sample conditions}

\item{groupBy}{a string, or a vector of sample conditions indicating the primary effector to be tested}
}
\value{
a list including a data.frame of targets (sample factors) and a vector indicating which column in the targets is the primary factor
}
\description{
extracting targets and primary factor based on the imported arguments
}
\examples{
Block <- c(1,1,2,3,3,4,1,1,2,3,3,4)
Treat <- factor(rep(1:2, each=6))
Time <- rep(rep(1:2, each=3), 2)
nbarcodes <- 50
nsamples <- 12
count <- matrix(rnorm(nbarcodes*nsamples), nbarcodes, nsamples) \%>\% abs()
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- Barbie::createBarbie(count, data.frame(Treat=Treat, Time=Time))
extarctTargetsAndPrimaryFactor(Barbie = Barbie, groupBy = "Treat")
}
