% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarcodePairCorrelation.R
\name{extractBarcodePairs}
\alias{extractBarcodePairs}
\title{generate pair wise Barcode correlation and standardize known cluster information}
\usage{
extractBarcodePairs(Barbie, method = "pearson", BarcodeClusters = NULL)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{method}{a string value to choose correlation method, default "pearson"}

\item{BarcodeClusters}{a list of known groups containing different Barcodes or a vector/array indicating Barcode groups}
}
\value{
an updated Barbie object including Barcode group information
}
\description{
generate pair wise Barcode correlation and standardize known cluster information
}
\examples{
nbarcodes <- 50
nsamples <- 12
count <- matrix(rnorm(nbarcodes*nsamples), nbarcodes, nsamples) \%>\% abs()
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- createBarbie(count)
Barbie:::extractBarcodePairs(Barbie, BarcodeClusters=c(rep(1:10, 5)))
}
