% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractTargetsAndPrimaryFactor}
\alias{extractTargetsAndPrimaryFactor}
\title{extracting targets and primary factor}
\usage{
extractTargetsAndPrimaryFactor(Barbie, targets = NULL, sampleGroups = NULL)
}
\arguments{
\item{Barbie}{a \code{Barbie} object created by \code{createBarbie()}}

\item{targets}{a matrix or data.frame storing sample conditions}

\item{sampleGroups}{a string, or a vector of sample conditions
indicating the primary effector to be tested}
}
\value{
a list including a data.frame of targets (sample factors)
and a vector indicating which column in the targets is the primary factor
}
\description{
\code{extractTargetsAndPrimaryFactor} extracts targets data.frame
and points to primary factor based on the imported arguments
}
\examples{
Treat <- factor(rep(1:2, each=6))
Time <- rep(rep(1:2, each=3), 2)
nbarcodes <- 50
nsamples <- 12
count <- matrix(rnorm(nbarcodes*nsamples), nbarcodes, nsamples) \%>\% abs()
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- createBarbie(count, data.frame(Treat=Treat, Time=Time))
Barbie:::extarctTargetsAndPrimaryFactor(
  Barbie = Barbie, sampleGroups = "Treat")
}
