% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarcodePairCorrelation.R
\name{inspectCorrelatingClusters}
\alias{inspectCorrelatingClusters}
\title{After running \code{clusterCorrelatingBarcodes}, inspect the barcode clusters,
the cluster sizes, and barcode CPM within each cluster.}
\usage{
inspectCorrelatingClusters(barbieQ)
}
\arguments{
\item{barbieQ}{A \code{barbieQ} object created by the \link{createBarbieQ} function,
then updated by the \link{clusterCorrelatingBarcodes} function.}
}
\value{
A list containing three ggplot objects, each for visualizing
the cluster structures, number of barcodes per cluster, and barcode CPM per cluster.
}
\description{
\code{inspectCorrelatingClusters()} visualize the cluster details generated by
calling \code{clusterCorrelatingBarcodes()}.
}
\examples{
nbarcodes <- 50
nsamples <- 12
count <- abs(matrix(rnorm(nbarcodes * nsamples), nbarcodes, nsamples))
rownames(count) <- paste0('Barcode', seq_len(nbarcodes))
barbieQ <- createBarbieQ(count)
clusterCorrelatingBarcodes(barbieQ, preDefinedCluster = list(
 group1 = c('Barcode1', 'Barcode2', 'Barcode3'), group2 = c('Barcode4', 'Barcode5')))
inspectCorrelatingClusters(barbieQ)
}
