% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{mergeLists}
\alias{mergeLists}
\title{Title Merging several lists into one list and include all elements in the old lists}
\usage{
mergeLists(...)
}
\arguments{
\item{...}{a list of lists}
}
\value{
a list
}
\description{
Title Merging several lists into one list and include all elements in the old lists
}
\examples{
list1 <- list(a = c("a"="#FFFFFF", "b" = "#CCCCFF"),
              b = c("x"="#000099"))
list2 <- list(a = c("a" = "#FFFFFF", "c"="#000000"),
              c = c("u" = "#CCCCCC"))
list3 <- list(b = c("x"="#999999"))
mergeLists(list1, list2, list3)
}
