% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{mergeMetadata}
\alias{mergeMetadata}
\title{Title merging several Barbie$metadata}
\usage{
mergeMetadata(...)
}
\arguments{
\item{...}{a list of metadata}
}
\value{
a data.frame
}
\description{
Title merging several Barbie$metadata
}
\examples{
mergeMetadata(Barbie1$metadata, Barbie2$metadata)
}
