% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarbieHeatmap.R
\name{plotBarbieHeatmap}
\alias{plotBarbieHeatmap}
\title{plotting Barcode CPM (or occurrence) in a Heatmap with sample annotations}
\usage{
plotBarbieHeatmap(
  Barbie,
  value = "CPM",
  splitSamples = FALSE,
  targets = NULL,
  sampleGroups = NULL,
  barcodeAnnotation = NULL,
  sampleAnnotation = NULL
)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{value}{a string value choosing to present "CPM" or "occurrence"}

\item{splitSamples}{a logical value deciding whether to split samples into slices}

\item{targets}{a data.frame containing sample conditions - each effector is a column}

\item{sampleGroups}{a vector or a string value indicating the primary effector in targets}

\item{barcodeAnnotation}{a row Annotation object created by ComplexHeatmap::rowAnnotation()}

\item{sampleAnnotation}{a column Annotation object created by ComplexHeatmap::HeatmapAnnotation()}
}
\value{
a "Heatmap" S4 object
}
\description{
plotting Barcode CPM (or occurrence) in a Heatmap with sample annotations
}
\examples{
HSC <- Barbie::HSC
plotBarbieHeatmap(Barbie = HSC)
}
