% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeBias.R
\name{plotBarcodeBiasHeatmap}
\alias{plotBarcodeBiasHeatmap}
\title{Plot Barcodes categorized by significant change using a Heatmap}
\usage{
plotBarcodeBiasHeatmap(
  Barbie,
  value = "CPM",
  elementName = NULL,
  sampleAnnotation = NULL
)
}
\arguments{
\item{Barbie}{A \code{Barbie} object created by the \link{createBarbie} function,
updated with Barcode test results by calling the \link{testBarcodeBias} function.}

\item{value}{A string indicating what to visualize.
Defaults to "CPM". Options include: "CPM" and "occurrence".}

\item{elementName}{A string indicating name of the test
conducted and stored in \code{Barbie$testBarcode}. Default to the last
test conducted and stored.}

\item{sampleAnnotation}{A column Annotation object created by the
\link[ComplexHeatmap:HeatmapAnnotation]{ComplexHeatmap::HeatmapAnnotation} function. Defaults to samples annotated
by the groups to be compared.}
}
\value{
A \code{Heatmap} S4 class object displaying the significance level
of Barcodes in Heatmap annotations.
}
\description{
\code{plotBarcodeBiasHeatmap()} uses the Heatmap annotations to visualize the
significance level of each Barcode in differential proportion or occurrence,
as determined by the \link{testBarcodeBias} function.
}
\examples{
Block <- c(1, 1, 2, 3, 3, 4, 1, 1, 2, 3, 3, 4)
Treat <- factor(rep(c("ctrl", "drug"), each = 6))
Time <- rep(rep(1:2, each = 3), 2)
nbarcodes <- 50
nsamples <- 12
count <- abs(matrix(rnorm(nbarcodes * nsamples), nbarcodes, nsamples))
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- createBarbie(count, data.frame(Treat = Treat, Time = Time))
testBB <- testBarcodeBias(Barbie, sampleGroups = "Treat")
plotBarcodeBiasHeatmap(Barbie = testBB, elementName = "diffProp_Treat")
}
