% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeBias.R
\name{plotBarcodeBiasScatterPlot}
\alias{plotBarcodeBiasScatterPlot}
\title{Plot Barcodes catogorized by significant change using a dot plot}
\usage{
plotBarcodeBiasScatterPlot(
  Barbie,
  elementName = NULL,
  reorderRank = FALSE,
  pValuesAdjusted = TRUE,
  xAxis = "avgRank"
)
}
\arguments{
\item{Barbie}{A \code{Barbie} object created by the \link{createBarbie} function,
updated with Barcode test results by calling the \link{testBarcodeBias} function.}

\item{elementName}{A string indicating name of the test
conducted and stored in \code{Barbie$testBarcode}. Default to the last
test conducted and stored.}

\item{reorderRank}{A logical value deciding whether to reorder
Barcode ranks within each sample. Defaults to FALSE.}

\item{pValuesAdjusted}{A logical value indicating which p.value to present.
TRUE to present adjusted p.values. FALSE to present original p.values.
Defaults to TRUE.}

\item{xAxis}{A string indicating what to visualise on the x scale of the
dot plot. Options include: "avgRank", "totalOcc", "avgLogCPM",
and "avgProportion". Defaults to "avgRank", representing the average
rank of Barcodes across samples.}
}
\value{
A \code{ggplot} S3 class object displaying the significance level
against other properties of Barcodes in a dot plot.
}
\description{
\code{plotBarcodeBiasScatterPlot()} use a dot plot to visualize the significance
level of each Barcode in differential proportion or occurrence,
as determined by the \link{testBarcodeBias} function.
P.values are plotted against other properties of Barcodes
specified by \code{xAxis}.
}
\examples{
Block <- c(1, 1, 2, 3, 3, 4, 1, 1, 2, 3, 3, 4)
Treat <- factor(rep(c("ctrl", "drug"), each = 6))
Time <- rep(rep(1:2, each = 3), 2)
nbarcodes <- 50
nsamples <- 12
count <- abs(matrix(rnorm(nbarcodes * nsamples), nbarcodes, nsamples))
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- createBarbie(count, data.frame(Treat = Treat, Time = Time))
testBB <- testBarcodeBias(Barbie, sampleGroups = "Treat")
plotBarcodeBiasScatterPlot(Barbie = testBB, elementName = "diffProp_Treat")
}
