% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeSankey.R
\name{plotBarcodeSankey}
\alias{plotBarcodeSankey}
\title{plotting relative Barcode total contribution in Sankey plot - top vs bottom Barcodes}
\usage{
plotBarcodeSankey(Barbie)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}
}
\value{
a "ggplot" S3 class object
}
\description{
plotting relative Barcode total contribution in Sankey plot - top vs bottom Barcodes
}
\examples{
HSC <- Barbie::HSC
plotBarcodeSankey(HSC)
}
