% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignifBarcodeProportion.R
\name{plotSignifBarcodeProportion}
\alias{plotSignifBarcodeProportion}
\title{Plot Barcode contributions as mean Barcode proportion across samples}
\usage{
plotSignifBarcodeProportion(barbieQ)
}
\arguments{
\item{barbieQ}{A \code{SummarizedExperiment} object created by the \link{createBarbieQ} function,
updated with Barcode test results by calling the \link{testBarcodeBias} function.}
}
\value{
A \code{ggplot} S3 class object displaying Barcode contributions in a
bar plot.
}
\description{
\code{plotBarcodeProportion()} visualizes the overall proportion of Barcodes
within groups of significance, as determined by \link{testBarcodeBias}, for each sample.
}
\examples{
## sample conditions and color palettes
sampleConditions <- data.frame(
  Treat = factor(rep(c('ctrl', 'drug'), each = 6)),
  Time = rep(rep(seq_len(2), each = 3), 2)
)
conditionColor <- list(
  Treat = c(ctrl = '#999999', drug = '#112233'),
  Time = c('1' = '#778899', '2' = '#998877')
)
## Barcode count data
nbarcodes <- 50
nsamples <- 12
barcodeCount <- abs(matrix(10, nbarcodes, nsamples))
barcodeCount[seq(21, 50), ] <- 0.0001
rownames(barcodeCount) <- paste0('Barcode', seq_len(nbarcodes))
## create a `barbieQ` object
myBarbieQ <- createBarbieQ(barcodeCount, sampleConditions, conditionColor)
plotSignifBarcodeProportion(myBarbieQ)
}
