% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{returnNumMat}
\alias{returnNumMat}
\title{extracting a numeric matrix out of the input object}
\usage{
returnNumMat(object)
}
\arguments{
\item{object}{a matrix, a data.frame, or a vector}
}
\value{
a numeric matrix with NAs retained
}
\description{
extracting a numeric matrix out of the input object
}
\examples{
mat <- data.frame(id = letters[1:5], matrix(1:25,5,5))
Barbie:::returnNumMat(mat)
}
