% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyeasy.R
\name{showName}
\alias{showName}
\title{print names checking building env}
\usage{
showName(myname)
}
\arguments{
\item{myname}{a string}
}
\value{
a string
}
\description{
print names checking building env
}
\examples{
showName("Lily!")
showName("second exmp.")
}
