% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{subsetSamplesByMetadata}
\alias{subsetSamplesByMetadata}
\title{Title make a subset of columns from all components of a 'Barbie' object based on specified conditions in the metadata}
\usage{
subsetSamplesByMetadata(Barbie, factor, specifiedConditions, keep = TRUE)
}
\arguments{
\item{Barbie}{an object created by the 'createBarbie()' function in the 'Barbie' package}

\item{factor}{a string specifying the factor name in 'Barbie$metadata'}

\item{specifiedConditions}{a string or a string vector specifying the conditions in the specified factor}

\item{keep}{a logical value indicating whether to retain or exclude the columns (samples) of the specified conditions}
}
\value{
an updated Barbie object with subsets of its components
}
\description{
Title make a subset of columns from all components of a 'Barbie' object based on specified conditions in the metadata
}
\examples{
myBarbie <- createBarbie(
  object=matrix(c(1:20), nrow=5, ncol=4),
  target=data.frame(fac=c(1, 1, 2, 2)))
subsetColumnsByMetadata(Barbie=myBarbie, factor="fac", specifiedConditions=1, keep=TRUE)
}
