% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagTopBarcodes.R
\name{tagTopBarcodes}
\alias{tagTopBarcodes}
\title{Title}
\usage{
tagTopBarcodes(
  Barbie,
  activeColumns = NULL,
  threshold = 0.99,
  minTopAppearance = 1
)
}
\arguments{
\item{Barbie}{an object created by the 'createBarbie()' function in the 'Barbie' package}

\item{activeColumns}{a logical vector indicating which column (sample) should be considered when determining the 'top Barcodes'}

\item{threshold}{a number ranging from 0 to 1 used for thresholding the Barcodes' contribution}

\item{minTopAppearance}{An integer indicating the minimum times a barcode must be 'top' in each column to be considered a 'top barcode' across samples}
}
\value{
a 'Barbie' obejct updated with isTop information
}
\description{
Title
}
\examples{
myBarbie <- createBarbie(mymat)
myBarbie <- tagTopBarcodes(myBarbie)
}
