% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDiffOcc.R
\name{testDiffOcc}
\alias{testDiffOcc}
\title{differential occurrence test on Barcodes across sample groups}
\usage{
testDiffOcc(
  Barbie,
  regularization = "firth",
  mycontrasts = NULL,
  contrastLevels = NULL,
  designMatrix = NULL
)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{regularization}{a string, options: "firth" and "none", choosing regularization method}

\item{mycontrasts}{a contast matrix generated by limma::makeContrasts}

\item{contrastLevels}{a charactor vector indicating the levels in specified groupBy}

\item{designMatrix}{a numeric matrix indicating sample conditions}
}
\value{
an updated Barbie object storing test results
}
\description{
differential occurrence test on Barcodes across sample groups
}
\examples{
Barbie <- testDiffOcc(Barbie, regularization="firth",
mycontrasts = mycontrasts, contrastLevels = contrastLevels,
designMatrix = designMatrix)
}
