% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDiffProp.R
\name{testDiffProp}
\alias{testDiffProp}
\title{differential proportion test on Barcodes across sample groups}
\usage{
testDiffProp(
  Barbie,
  mycontrasts = NULL,
  contrastLevels = NULL,
  designMatrix = NULL,
  block = NULL
)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{mycontrasts}{a contast matrix generated by limma::makeContrasts}

\item{contrastLevels}{a charactor vector indicating the levels in specified groupBy}

\item{designMatrix}{a numeric matrix indicating sample conditions}

\item{block}{a vector of indicating sample duplicates}
}
\value{
an updated Barbie object storing test results
}
\description{
differential proportion test on Barcodes across sample groups
}
\examples{
Barbie <- testDiffProp(Barbie = Barbie, mycontrasts = mycontrasts,
                       contrastLevels = contrastLevels, designMatrix = designMatrix,
                       block = block)
}
