% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkBarbieDimensions}
\alias{checkBarbieDimensions}
\title{check whether components of Barbie object have identical sample dimensions and Barcode dimensions.}
\usage{
checkBarbieDimensions(Barbie)
}
\arguments{
\item{Barbie}{'Barbie' object created by createBarbie()}
}
\value{
a logical value
}
\description{
check whether components of Barbie object have identical sample dimensions and Barcode dimensions.
}
\examples{
Treat <- factor(rep(1:2, each=6))
Time <- rep(rep(1:2, each=3), 2)
nbarcodes <- 50
nsamples <- 12
count <- abs(matrix(rnorm(nbarcodes*nsamples), nbarcodes, nsamples))
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- Barbie::createBarbie(count, data.frame(Treat=Treat, Time=Time))
Barbie:::checkBarbieDimensions(Barbie)
}
