% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarcodePairCorrelation.R
\name{clusterCorrelatingBarcodes}
\alias{clusterCorrelatingBarcodes}
\title{cluster Barcodes based on pair wise correlation}
\usage{
clusterCorrelatingBarcodes(
  Barbie,
  method = "pearson",
  corCutoff = 0.95,
  dataCutoff = 0,
  BarcodeClusters = NULL
)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{method}{a string value to choose correlation method, default "pearson"}

\item{corCutoff}{a numeric value to decide high correlation level, default 0.95}

\item{dataCutoff}{a numeric value to decide the minimum level of log2(mean CPM)}

\item{BarcodeClusters}{a list of known groups containing different Barcodes or a vector/array indicating Barcode groups}
}
\value{
a ggplot S3 object
}
\description{
cluster Barcodes based on pair wise correlation
}
\examples{
nbarcodes <- 50
nsamples <- 12
count <- abs(matrix(rnorm(nbarcodes*nsamples), nbarcodes, nsamples))
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- Barbie::createBarbie(count)
clusterCorrelatingBarcodes(Barbie, BarcodeClusters=c(rep(1:10, 5)))
}
