% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{combineBarbies}
\alias{combineBarbies}
\title{Title combining several 'Barbie' object by binding matrix components by samples and merging list components}
\usage{
combineBarbies(...)
}
\arguments{
\item{...}{a list of 'Barbie' object}
}
\value{
a 'Barbie' object
}
\description{
Title combining several 'Barbie' object by binding matrix components by samples and merging list components
}
\examples{
combineBarbies(Barbie1, Barbie2, Barbie3)
}
