% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodePairCorrelation.R
\name{inspectCorrelatingBarcodes}
\alias{inspectCorrelatingBarcodes}
\title{Visualising the barcode clusters, the cluster sizes, and barcode proportion
within each cluster, after running the \code{clusterCorrelatingBarcodes} function.}
\usage{
inspectCorrelatingBarcodes(barbieQ)
}
\arguments{
\item{barbieQ}{A \code{barbieQ} object with information of clusters
of correlating barcodes, which should be predicted by the
\link{clusterCorrelatingBarcodes} function, after initialized the barbieQ object
by the \link{createBarbieQ} function.}
}
\value{
A list containing three ggplot objects, each for visualizing
the cluster structures, number of barcodes per cluster, and barcode proportion per cluster.
}
\description{
\code{inspectCorrelatingBarcodes()} visualize the cluster details generated by
calling \code{clusterCorrelatingBarcodes()}.
}
\examples{
nbarcodes <- 50
nsamples <- 12
count <- abs(matrix(rnorm(nbarcodes * nsamples), nbarcodes, nsamples))
rownames(count) <- paste0('Barcode', seq_len(nbarcodes))
barbieQ <- createBarbieQ(count)
barbieQ <- clusterCorrelatingBarcodes(barbieQ, preDefinedCluster = list(
 group1 = c('Barcode1', 'Barcode2', 'Barcode3'), group2 = c('Barcode4', 'Barcode5')))
inspectCorrelatingBarcodes(barbieQ)
}
