% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodePairCorrelation.R
\name{mergeCorrelatingBarcodes}
\alias{mergeCorrelatingBarcodes}
\title{Merging correlating barcodes, by keeping a representative barcode
and removing the rest for each cluster. The representative barcode is
the one with the highest mean proportion across samples.}
\usage{
mergeCorrelatingBarcodes(barbieQ_clustered, barbieQ_toMerge = NULL)
}
\arguments{
\item{barbieQ_clustered}{A \code{barbieQ} object with information of clusters
of correlating barcodes, which should be predicted by the
\link{clusterCorrelatingBarcodes} function, after initialized the barbieQ object
by the \link{createBarbieQ} function.}

\item{barbieQ_toMerge}{A \code{barbieQ} object, whose barcodes will be merged
based on clusters specified by the \code{barbieQ_clustered} parameter.
If not specified, defaults to the same \code{barbieQ} object as \code{barbieQ_clustered}.}
}
\value{
A \code{barbieQ} object updated by merging the correlating barcodes into one
by selecting a representative from each cluster and deleting the rest.
The processing information saved in \code{metadata$predicted_barcode_clusters} and
\code{metadata$removed_barcodes}.
}
\description{
Merging correlating barcodes, by keeping a representative barcode
and removing the rest for each cluster. The representative barcode is
the one with the highest mean proportion across samples.
}
\examples{
nbarcodes <- 50
nsamples <- 12
count <- abs(matrix(rnorm(nbarcodes * nsamples), nbarcodes, nsamples))
rownames(count) <- paste0('Barcode', seq_len(nbarcodes))
barbieQ <- createBarbieQ(count)
barbieQ <- clusterCorrelatingBarcodes(barbieQ, preDefinedCluster = list(
 group1 = c('Barcode1', 'Barcode2', 'Barcode3'), group2 = c('Barcode4', 'Barcode5')))
mergeCorrelatingBarcodes(barbieQ, barbieQ)
}
