% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeBias.R
\name{plotBarcodeBiasHeatmap}
\alias{plotBarcodeBiasHeatmap}
\title{plotting test results of Barcodes by Heatmap}
\usage{
plotBarcodeBiasHeatmap(
  Barbie,
  value = "CPM",
  elementName = NULL,
  sampleAnnotation = NULL
)
}
\arguments{
\item{Barbie}{Barbie object created by createBarbie(), with test results stored by calling testBarcodeBias()}

\item{value}{a string value choosing to present "CPM" or "occurrence"}

\item{elementName}{a string value indicating the test being conducted and stored in Barbie$testBarcode}

\item{sampleAnnotation}{a column Annotation object created by ComplexHeatmap::HeatmapAnnotation()}
}
\value{
a "Heatmap" S4 object
}
\description{
plotting test results of Barcodes by Heatmap
}
\examples{
HSC <- Barbie::HSC
HSC <- testBarcodeBias(HSC , groupBy = "treat", contrastLevels = c("IV", "IT"))
plotBarcodeBiasHeatmap(HSC)
}
