% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeBias.R
\name{plotBarcodeBiasScatterPlot}
\alias{plotBarcodeBiasScatterPlot}
\title{plotting test results of Barcodes by scatter plot}
\usage{
plotBarcodeBiasScatterPlot(
  Barbie,
  elementName = NULL,
  reorderRank = FALSE,
  pValuesAdjusted = TRUE,
  xAxis = "avgRank"
)
}
\arguments{
\item{Barbie}{Barbie object created by createBarbie(), with test results stored by calling testBarcodeBias()}

\item{elementName}{a string value indicating the test being conducted and stored in Barbie$testBarcode}

\item{reorderRank}{a logical value deciding whether to reorder Barcode rank within each sample}

\item{pValuesAdjusted}{a logical value choosing to present adjusted p.values or original p.values}

\item{xAxis}{a string value indicating what to visualise on x scale}
}
\value{
a "ggplot" S3 class object
}
\description{
plotting test results of Barcodes by scatter plot
}
\examples{
HSC <- Barbie::HSC
BB <- createBarbie(object = HSC$assay, target = HSC$metadata)
rownames(BB$metadata) <- BB$metadata[,1]
BB$metadata <- BB$metadata[,-1]
testBB <- testBarcodeBias(BB, groupBy = "treat", contrastLevels = c("IV", "IT"))

plotBarcodeBiasScatterPlot(Barbie = testBB)
}
