% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarcodePairCorrelation.R
\name{plotBarcodePairCorrelation}
\alias{plotBarcodePairCorrelation}
\title{plot Barcode pair wise correlation}
\usage{
plotBarcodePairCorrelation(
  Barbie,
  method = "pearson",
  dataVisual = "mean",
  corCutoff = 0.95,
  dataCutoff = 0,
  BarcodeClusters = NULL
)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{method}{a string value to choose correlation method, default "pearson"}

\item{dataVisual}{a sring value to choose visualization data from "mean" or "max" of Barcode CPM in each pair}

\item{corCutoff}{a numeric value to decide high correlation level, default 0.95}

\item{dataCutoff}{a numeric value to decide the minimum level of log2(mean CPM) to make correlation level convincing}

\item{BarcodeClusters}{a list of known groups containing different Barcodes or a vector/array indicating Barcode groups}
}
\value{
a list of two data.frame including correlation of all pairs of Barcodes and known correlated Barcode names
}
\description{
plot Barcode pair wise correlation
}
\examples{
nbarcodes <- 50
nsamples <- 12
count <- matrix(rnorm(nbarcodes*nsamples), nbarcodes, nsamples) \%>\% abs()
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- Barbie::createBarbie(count)
plotBarcodePairCorrelation(Barbie, BarcodeClusters=c(rep(1:10, 5)))
}
