% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodePareto.R
\name{plotBarcodePareto}
\alias{plotBarcodePareto}
\title{plotting Barcode proportion by a circular bar plot}
\usage{
plotBarcodePareto(Barbie)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}
}
\value{
a "ggplot" S3 class object
}
\description{
plotting Barcode proportion by a circular bar plot
}
\examples{
HSC <- Barbie::HSC
plotBarcodePareto(HSC)
}
