% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarcodeProportion.R
\name{plotBarcodeProportion}
\alias{plotBarcodeProportion}
\title{plotting Barcode contributions (percentage of average Barcode proportion across samples)}
\usage{
plotBarcodeProportion(Barbie, coordFixed = FALSE, colorGradient = FALSE)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{coordFixed}{a logical value to decide whether to coordinate x and y scale}

\item{colorGradient}{a logical value to choose bar colors being consistant or gradient}
}
\value{
a "ggplot" S3 class object
}
\description{
plotting Barcode contributions (percentage of average Barcode proportion across samples)
}
\examples{
HSC <- Barbie::HSC
plotBarcodeProportion(HSC)
plotBarcodeProportion(HSC, colorGradient=TRUE)
}
