% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSamplePairCorrelation.R
\name{plotSamplePairCorrelation}
\alias{plotSamplePairCorrelation}
\title{plotting sample pair-wise correlation in Heatmap}
\usage{
plotSamplePairCorrelation(
  Barbie,
  sampleOrder = NULL,
  targets = NULL,
  groupBy = NULL,
  method = "pearson"
)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{sampleOrder}{a charactor vector indicating samples ordered by which factor in targets}

\item{targets}{a data.frame containing sample conditions - each effector is a column}

\item{groupBy}{a vector or a string value indicating the primary effector in targets}

\item{method}{a string value choose correlation method from "pearson" or "spearman"}
}
\value{
a "Heatmap" S4 obejct
}
\description{
plotting sample pair-wise correlation in Heatmap
}
\examples{
HSC <- Barbie::HSC
plotSamplePairCorrelation(Barbie = HSC)

HSC$metadata$treat <- factor(HSC$metadata$treat, levels = c("IF", "IT", "IV"))
HSC$metadata$lineage <- factor(HSC$metadata$lineage, levels = c("Bcell", "Tcell", "immature", "Myeloid"))
sampleOrder <- c("time", "treat", "lineage", "celltype", "tissue", "mouse")
plotSamplePairCorrelation(Barbie = HSC, sampleOrder = sampleOrder)
}
