% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{subsetBarcodes}
\alias{subsetBarcodes}
\title{Title make a subset of rows from all components of a 'Barbie' object based on specified rows}
\usage{
subsetBarcodes(Barbie, retainedRows)
}
\arguments{
\item{retainedRows}{a logical vector indicating each row to be retained or excluded}

\item{Barbiean}{object created by the 'createBarbie()' function in the 'Barbie' package}
}
\value{
an updated Barbie object with subsets of its components
}
\description{
Title make a subset of rows from all components of a 'Barbie' object based on specified rows
}
\examples{
myBarbie <- createBarbie(
  object=matrix(c(1:20), nrow=5, ncol=4),
  target=data.frame(fac=c(1, 1, 2, 2)))
subsetBarcodes(Barbie=myBarbie, retainedRows=c(TRUE, TRUE, FALSE, FALSE, FALSE))
}
