% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{subsetSamples}
\alias{subsetSamples}
\title{Title make a subset of columns from all components of a 'Barbie' object based on specified columns}
\usage{
subsetSamples(Barbie, retainedColumns)
}
\arguments{
\item{Barbie}{an object created by the 'createBarbie()' function in the 'Barbie' package}

\item{retainedColumns}{a logical vector indicating each column to be retained or excluded}
}
\value{
an updated Barbie object with subsets of its components
}
\description{
Title make a subset of columns from all components of a 'Barbie' object based on specified columns
}
\examples{
myBarbie <- createBarbie(
  object=matrix(c(1:20), nrow=5, ncol=4),
  target=data.frame(fac=c(1, 1, 2, 2)))
subsetColumns(Barbie=myBarbie, retainedColumns=c(TRUE, TRUE, FALSE, FALSE))
}
