% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifierMethods.R
\name{subsetSamplesByMetadata}
\alias{subsetSamplesByMetadata}
\title{Trim samples of a \code{barbieQ} object based on specified specified
sample conditions}
\usage{
subsetSamplesByMetadata(
  barbieQ,
  factor = NULL,
  specifiedConditions = NULL,
  keep = TRUE
)
}
\arguments{
\item{barbieQ}{A \code{barbieQ} object created by the \link{createBarbieQ} function.}

\item{factor}{A string representing the name of a factor in
\code{barbieQ$metadata}. Defaults to the first factor in \code{barbieQ$metadata}.}

\item{specifiedConditions}{A string vector specifying the conditions
in the specified factor. Defaults to all conditions in the specified factor.}

\item{keep}{A logical value: TRUE to retain or FALSE to exclude
the columns (samples) of the specified conditions. Defaults to TRUE.}
}
\value{
A \code{barbieQ} object with:
\itemize{
\item Updated components reflecting the subsetted samples, including \code{assay},
\code{metadata}, \code{proportion}, \code{CPM}, \code{occurrence}, \code{rank}, and \code{isTop$mat}.
\item Other components are inherited from the original \code{barbieQ} object.
}
}
\description{
In the \code{matrix} and \code{data.frame} components of a \code{barbieQ} object,
samples are organized in columns and Barcodes in rows. The \code{metadata}
component in \code{barbieQ} is a \code{data.frame} representing sample conditions,
organized by various experimental factors.
\code{subsetSamplesByMetadata()} trims the columns based on specified
\code{specifiedConditions} in the specified \code{factor} in \code{barbieQ$metadata}.
}
\examples{
myBarbieQ <- createBarbieQ(
  object = matrix(seq_len(20), nrow = 5, ncol = 4),
  target = data.frame(fac = c(1, 1, 2, 2))
)
subsetSamplesByMetadata(
  barbieQ = myBarbieQ, factor = "fac", specifiedConditions = 1, keep = TRUE
)
}
