% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagTopBarcodes.R
\name{tagTopBarcodes}
\alias{tagTopBarcodes}
\title{Tag each Barcode as being part of the major contributors or not}
\usage{
tagTopBarcodes(
  Barbie,
  activeSamples = NULL,
  proportionThreshold = 0.99,
  minTopFrequency = 1
)
}
\arguments{
\item{Barbie}{An object created by the \link{createBarbie} function.}

\item{activeSamples}{a logical vector indicating which samples (columns)
to consider when determining the \emph{top} Barcodes across the entire
dataset. Default to considering all samples in the \code{Barbie} object.}

\item{proportionThreshold}{A numeric value ranging from 0 to 1,
used as a threshold for determining \emph{top} Barcodes in each sample
based on their combined proportion in theat sample. Default to 0.99.}

\item{minTopFrequency}{An integer specifying the minimum number of times
a Barcode must be tagged as \emph{top} across the selected samples
(specified by \code{activeSamples}) in order to be considered \emph{top}
for the entire dataset. Default to 1.}
}
\value{
A \code{Barbie} object updating \code{isTop} while inheriting other components
from the \code{Barbie} object passed into the function.
See \code{Barbie} structure in \link{createBarbie}.
\code{isTop} is a list including:
\itemize{
\item \code{vec}: a logical vector tagging Barcodes as \emph{top}
across the entire dataset.
\item \code{mat}: a logical matrix tagging Barcodes as \emph{top} in each sample.
}
}
\description{
\code{tagTopBarcodes()} is designed for better filtering out the background noise,
i.e., Barcodes that consistently have low contributions across samples.
Each Barcode is tagged as one of the major contributing Barcodes
(\emph{top} Barcodes) or not in each sample, based on whether their
combined proportion passes a defined threshold in the sample.
Across the entire dataset, a Barcode is considered \emph{top}
if it is tagged as \emph{top} in a number of samples,
meeting a specified appearance threshold across all selected samples.
}
\examples{
## create a `Barbie` object
## sample conditions and color palettes
sampleConditions <- data.frame(
  Treat = factor(rep(c("ctrl", "drug"), each = 6)),
  Time = rep(rep(1:2, each = 3), 2)
)
conditionColor <- list(
  Treat = c(ctrl = "#999999", drug = "#112233"),
  Time = c("1" = "#778899", "2" = "#998877")
)
## Barcode count data
nbarcodes <- 50
nsamples <- 12
barcodeCount <- abs(matrix(rnorm(nbarcodes * nsamples), nbarcodes, nsamples))
rownames(barcodeCount) <- paste0("Barcode", 1:nbarcodes)
## create a `Barbie` object
myBarbie <- createBarbie(barcodeCount, sampleConditions, conditionColor)
## tag top Barcodes
tagTopBarcodes(myBarbie)
}
