% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagTopBarcodes.R
\name{tagTopEachColumn}
\alias{tagTopEachColumn}
\title{Tag top Barcodes in each column (sample).}
\usage{
tagTopEachColumn(tempCol, threshold = 0.99)
}
\arguments{
\item{tempCol}{a numeric vector of Barcode counts in a sample, usually being a column in a count matrix}

\item{threshold}{a number ranging from 0 to 1 used for thresholding the Barcodes' contribution}
}
\value{
a logical vector indicating whether each Barcode is considered as 'top' in this vector
}
\description{
Tag top Barcodes in each column (sample).
}
\examples{
myCount <- c(1, 2, 3, 98, NA, NA)
tagTopEachColumn(myCount, 0.99)
}
