% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testBarcodeBias.R
\name{testBarcodeBias}
\alias{testBarcodeBias}
\title{Test Barcodes bias across sample groups}
\usage{
testBarcodeBias(
  Barbie,
  method = "diffProp",
  targets = NULL,
  groupBy = NULL,
  contrastLevels = NULL,
  designFormula = NULL,
  designMatrix = NULL,
  block = NULL,
  transformation = "asin-sqrt",
  regularization = "firth"
)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{method}{a string, options: "diffProp" and "diffOcc", choosing test method}

\item{targets}{a matrix or data.frame storing sample conditions}

\item{groupBy}{a string, or a vector of sample conditions indicating the primary effector to be tested}

\item{contrastLevels}{a charactor vector indicating the levels in specified groupBy}

\item{designFormula}{a formula to compute the designMatrix}

\item{designMatrix}{a numaric matrix to standardize 'targets'}

\item{block}{a vector of indicating sample duplicates}

\item{regularization}{a string, options: "firth" and "none", choosing regularization method in "diffOcc" test}
}
\value{
an updated Barbie object storing test results
}
\description{
Test Barcodes bias across sample groups
}
\examples{
Block <- c(1,1,2,3,3,4,1,1,2,3,3,4)
Treat <- factor(rep(1:2, each=6))
Time <- rep(rep(1:2, each=3), 2)
nbarcodes <- 50
nsamples <- 12
count <- matrix(rnorm(nbarcodes*nsamples), nbarcodes, nsamples) \%>\% abs()
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- Barbie::createBarbie(count, data.frame(Treat=Treat, Time=Time))
}
