% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDiffProp.R
\name{testDiffProp}
\alias{testDiffProp}
\title{differential proportion test on Barcodes across sample groups}
\usage{
testDiffProp(
  Barbie,
  transformation = "asin-sqrt",
  mycontrasts = NULL,
  contrastLevels = NULL,
  designMatrix = NULL,
  block = NULL
)
}
\arguments{
\item{Barbie}{a Barbie object created by createBarbie()}

\item{mycontrasts}{a numeric vector generated by limma::makeContrasts}

\item{contrastLevels}{a character vector indicating the levels of the conditions in specified factor effector (primary predictor), note the conditions must be strings}

\item{designMatrix}{a numeric matrix indicating sample conditions}

\item{block}{a vector indicating sample duplicates}
}
\value{
an updated Barbie object storing test results
}
\description{
differential proportion test on Barcodes across sample groups
}
\examples{
Block <- c(1,1,2,3,3,4,1,1,2,3,3,4)
Treat <- factor(rep(c("ctrl", "drug"), each=6))
Time <- rep(rep(1:2, each=3), 2)
nbarcodes <- 50
nsamples <- 12
count <- matrix(rnorm(nbarcodes*nsamples), nbarcodes, nsamples) \%>\% abs()
rownames(count) <- paste0("Barcode", 1:nbarcodes)
Barbie <- Barbie::createBarbie(count, data.frame(Treat=Treat, Time=Time))
testDiffProp(
  Barbie = Barbie,
  mycontrasts = c(-1, 1, 0),
  contrastLevels = c("ctrl", "drug"),
  designMatrix = model.matrix(~0 + Treat + Time))
}
